refactor(database): 提取建表SQL到 schema.py，新增 force 参数和 data-refresh 入口 - 将建表/建索引逻辑从 __init__.py 抽取到独立的 schema.py - seed_data / compute_baseline / init_database 增加 force 参数，支持强制清空后重写 - Makefile 新增 data-refresh 目标，调用 init_database(force=True) - 修复 DB_PATH / _DATA_ROOT 路径层级错误（项目根目录 → backend/data）
