+++ backend/app/local_memory.py
@@ -16,7 +16,7 @@ from pathlib import Path
 
 logger = logging.getLogger("noterx.local_memory")
 
-_DATA_ROOT = Path(__file__).parent.parent.parent / "data"
+_DATA_ROOT = Path(__file__).parent.parent / "data"
 WORKSPACE_ROOT = os.path.join(_DATA_ROOT, "noterx_workspace")
 MEMORY_MD = os.path.join(WORKSPACE_ROOT, "MEMORY.md")
 MEMORY_DIR = os.path.join(WORKSPACE_ROOT, "memory")
